requirements.txt
speedtest-cli>=2.1.3
## synchronous gauge added in 1.23.0
opentelemetry-api>=1.23.0
azure-monitor-opentelemetry>=1.2.0
dnsdiag